chain, graph: Move BlockPtr into blockchain::types
