+++ graph/src/components/ethereum/types.rs
@@ -1,12 +1,8 @@
-use anyhow::anyhow;
 use ethabi::LogParam;
 use serde::{Deserialize, Serialize};
 use slog::{o, SendSyncRefUnwindSafeKV};
-use stable_hash::prelude::*;
-use stable_hash::utils::AsBytes;
+use std::sync::Arc;
 use std::{cmp::Ordering, convert::TryFrom};
-use std::{fmt, str::FromStr};
-use std::{fmt::Write, sync::Arc};
 use strum_macros::AsStaticStr;
 use web3::types::{
     Action, Address, Block, Bytes, Log, Res, Trace, Transaction, TransactionReceipt, H160, H256,
@@ -14,10 +10,10 @@ use web3::types::{
 };
 
 use crate::{
-    blockchain::BlockHash,
+    blockchain::BlockPtr,
     prelude::{
-        BlockNumber, CheapClone, DeploymentHash, EntityKey, MappingBlockHandler,
-        MappingCallHandler, MappingEventHandler, ToEntityKey,
+        BlockNumber, DeploymentHash, EntityKey, MappingBlockHandler, MappingCallHandler,
+        MappingEventHandler, ToEntityKey,
     },
 };
 
@@ -477,67 +473,6 @@ impl Clone for EthereumCallData {
     }
 }
 
-/// A block hash and block number from a specific Ethereum block.
-///
-/// Block numbers are signed 32 bit integers
-#[derive(Clone, Debug, PartialEq, Eq, Hash)]
-pub struct BlockPtr {
-    pub hash: BlockHash,
-    pub number: BlockNumber,
-}
-
-impl CheapClone for BlockPtr {}
-
-impl StableHash for BlockPtr {
-    fn stable_hash<H: StableHasher>(&self, mut sequence_number: H::Seq, state: &mut H) {
-        AsBytes(self.hash.0.as_ref()).stable_hash(sequence_number.next_child(), state);
-        self.number.stable_hash(sequence_number.next_child(), state);
-    }
-}
-
-impl BlockPtr {
-    /// Encodes the block hash into a hexadecimal string **without** a "0x" prefix.
-    /// Hashes are stored in the database in this format.
-    pub fn hash_hex(&self) -> String {
-        let mut s = String::with_capacity(self.hash.0.len() * 2);
-        for b in self.hash.0.iter() {
-            write!(s, "{:02x}", b).unwrap();
-        }
-        s
-    }
-
-    /// Block number to be passed into the store. Panics if it does not fit in an i32.
-    pub fn block_number(&self) -> BlockNumber {
-        self.number
-    }
-
-    pub fn hash_as_h256(&self) -> H256 {
-        H256::from_slice(self.hash_slice())
-    }
-
-    pub fn hash_slice(&self) -> &[u8] {
-        self.hash.0.as_ref()
-    }
-}
-
-impl fmt::Display for BlockPtr {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "#{} ({})", self.number, self.hash_hex())
-    }
-}
-
-impl<T> From<Block<T>> for BlockPtr {
-    fn from(b: Block<T>) -> BlockPtr {
-        BlockPtr::from((b.hash.unwrap(), b.number.unwrap().as_u64()))
-    }
-}
-
-impl<'a, T> From<&'a Block<T>> for BlockPtr {
-    fn from(b: &'a Block<T>) -> BlockPtr {
-        BlockPtr::from((b.hash.unwrap(), b.number.unwrap().as_u64()))
-    }
-}
-
 impl From<EthereumBlock> for BlockPtr {
     fn from(b: EthereumBlock) -> BlockPtr {
         BlockPtr::from((b.block.hash.unwrap(), b.block.number.unwrap().as_u64()))
@@ -547,69 +482,6 @@ impl From<EthereumBlock> for BlockPtr {
 impl<'a> From<&'a EthereumBlock> for BlockPtr {
     fn from(b: &'a EthereumBlock) -> BlockPtr {
         BlockPtr::from((b.block.hash.unwrap(), b.block.number.unwrap().as_u64()))
-    }
-}
-
-impl From<(Vec<u8>, i32)> for BlockPtr {
-    fn from((bytes, number): (Vec<u8>, i32)) -> Self {
-        BlockPtr {
-            hash: BlockHash::from(bytes),
-            number,
-        }
-    }
-}
-
-impl From<(H256, i32)> for BlockPtr {
-    fn from((hash, number): (H256, i32)) -> BlockPtr {
-        BlockPtr {
-            hash: hash.into(),
-            number,
-        }
-    }
-}
-
-impl From<(H256, u64)> for BlockPtr {
-    fn from((hash, number): (H256, u64)) -> BlockPtr {
-        let number = i32::try_from(number).unwrap();
-
-        BlockPtr::from((hash, number))
-    }
-}
-
-impl From<(H256, i64)> for BlockPtr {
-    fn from((hash, number): (H256, i64)) -> BlockPtr {
-        if number < 0 {
-            panic!("block number out of range: {}", number);
-        }
-
-        BlockPtr::from((hash, number as u64))
-    }
-}
-
-impl TryFrom<(&str, i64)> for BlockPtr {
-    type Error = anyhow::Error;
-
-    fn try_from((hash, number): (&str, i64)) -> Result<Self, Self::Error> {
-        let hash = hash.trim_start_matches("0x");
-        let hash = H256::from_str(hash)
-            .map_err(|e| anyhow!("Cannot parse H256 value from string `{}`: {}", hash, e))?;
-
-        Ok(BlockPtr::from((hash, number)))
-    }
-}
-
-impl TryFrom<(&[u8], i64)> for BlockPtr {
-    type Error = anyhow::Error;
-
-    fn try_from((bytes, number): (&[u8], i64)) -> Result<Self, Self::Error> {
-        let hash = if bytes.len() == H256::len_bytes() {
-            H256::from_slice(bytes)
-        } else {
-            return Err(anyhow!(
-                "invalid H256 value `{}` has {} bytes instead of {}"
-            ));
-        };
-        Ok(BlockPtr::from((hash, number)))
     }
 }
 
@@ -625,18 +497,6 @@ impl<'a> From<&'a BlockFinality> for BlockPtr {
             BlockFinality::Final(b) => b.into(),
             BlockFinality::NonFinal(b) => BlockPtr::from(&b.ethereum_block),
         }
-    }
-}
-
-impl From<BlockPtr> for H256 {
-    fn from(ptr: BlockPtr) -> Self {
-        ptr.hash_as_h256()
-    }
-}
-
-impl From<BlockPtr> for BlockNumber {
-    fn from(ptr: BlockPtr) -> Self {
-        ptr.number
     }
 }
