+++ graph/src/components/subgraph/proof_of_indexing/online.rs
@@ -3,7 +3,10 @@
 //! to the reference implementation, but this is updated incrementally
 
 use super::ProofOfIndexingEvent;
-use crate::prelude::{debug, BlockNumber, BlockPtr, DeploymentHash, Logger};
+use crate::{
+    blockchain::BlockPtr,
+    prelude::{debug, BlockNumber, DeploymentHash, Logger},
+};
 use lazy_static::lazy_static;
 use stable_hash::crypto::{Blake3SeqNo, SetHasher};
 use stable_hash::prelude::*;
